\name{my1rpackage}

\alias{-package}

\docType{package}

\title{
My first R package
}

\description{
  Contains a function for generating tikz plots and a dataset containing
  winter Olympics medal statistics by country.
}

\details{
It is a demo
}
}
\author{
 Zhendan Zhu
  
 Maintainer: Zhendan Zhu <zzhu11@jhu.edu>

}
\references{
  2012 Fall semester 550400 Homework 2 Description
}

\keyword{ package }

\examples{
require(my1rpack)

problem1('bmw1series', dnorm, c(-5, 5), 10, TRUE)

problem2
}