\name{problem2}

\alias{output}

\docType{data}

\title{
Winter Olympics: which country win the most medals?
}

\description{
  The data containing medal statistics by country.
}

\usage{data(problem2)}

\format{
  A list-mode object ith 45 slots, each of which contains the overall
  Gold, Silver, and Bronze medal counts:
  \describe{
    \item{\code{golds}}{Awesome!}
    \item{\code{silvers}}{Great!}
    \item{\code{bronze}}{Good!}
  }
}

\details{}

\source{\url{http://www.guardian.co.uk/news/datablog/2010/feb/11/winter-olympics-medals-by-country}}

\examples{
data(problem2)
}

\keyword{datasets}
