\name{my1rpack}
\alias{my1rpack-package}
\docType{package}
\title{My First R Package

}
\description{This R package contains 1 function and 1 dataset}


\details{This R package is the solution to problem3 of homework2 of class 550400

}

\author{Yu Du Email: ydu10@jhu.edu

}

\keyword{package}
\examples{
library(my1rpack)
}