\name{my1rpack}
\alias{-package}
\docType{package}
\title{R Package for HW 2}
\description{An R package built for homework 2 containing an R function for generating probability density functions and an organized data table of medal counts from the winter olympics 1924-2006}
\details{}
\author{Jonathan Ho <jho19@jhu.edu>}
\keyword{package}
\examples{}

