\name{my1rpack}
\alias{my1rpack-package}
\docType{package}
\title{Homework2}
\description{Includes problem1 and problem2 required in Homework 2 }
\details{ 
problem1: function that generates a .tex and .pdf file for the graph of the specified function over the specified domain
problem2: datasets that includes medal counts for Winter Olympics for different countries
}

\author{Joyce Tan jtan21@jhu.edu}

\keyword{package}

\examples{
problem1('dnorm',dnorm,c(-5,5), 100, T)
data(problem2)}

