\name{my1rpack}
\alias{my1rpack}

\docType{package}

\title{
  Homework 2
}

\description{
  There are a function for compiling plots and a database of medal count by countries.
}

\details{
\tabular{ll}{
Package: \tab my1rpack\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-11-01\cr
License: \tab GPL-3\cr
}
}
\author{
  Rong Fan <rfan4@jhu.edu>

}
\references{
  12Fall 550400 Homework2 Description
}

\keyword{ package }

\examples{
require(my1rpack)

problem1('bmw1series', dnorm, c(-5, 5), 10, TRUE)
data(problem2)
}
