\name{my1rpack-package}
\alias{my1rpack-package}
\alias{my1rpack}
\docType{package}
\title{
  Homework 2
}
\description{
  The package includes a function generating a plot for the specific function in a certain range and a function generating statistics of number of medals for each country
}
\details{
\tabular{ll}{
Package: \tab my1rpack\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-11-01\cr
License: \tab GPL-3\cr
}
}
\author{
  Shihong Li email<sli50@jhu.edu>
}
\references{
  2012 Fall semester 550400 Homework 2 Description
}
\keyword{ package }

\examples{
require(my1rpack)
problem1('bmw1series', dnorm, c(-5, 5), 10, TRUE)
problem2
}