\name{my1rpack}
\alias{my1rpack-package}
\docType{package}
\title{my1rpack}
\description{The package can generate Latex file to plot a function, and it provides a dataset grouping medals according to country. }
\details{The package consists of 2 parts: the 1st is a function which can generate Latex file to plot a function; the 2nd is a dataset which groups the number of Bronze, Gold and Silver Medals according to country. }
\author{Tianyu Luo }
\keyword{package}
\examples{ }