\name{my1rpack}
\alias{my1rpack-package}
\docType{package}
\title{My First R Package}
\description{This R package contains a function and a dataset. The function is defined in Problem 1 and the dataset is from Problem 2.}

\details{This R package is the solution to Problem 3 of homeworkset 2 of 550400}
\author{Xiaohan Yang email:xyang39@jhu.edu}

\keyword{package}
\examples{
package(my1rpack)
}
