\name{my1rpack}
\alias{-package}
\docType{package}
\title{Package to problems 1 and 2}
\description{This R package contains the solution to problems 1 and 2. }
\details{Problem 1 seeks to build and document an R function, which is the graphing function. Problem 2 seeks to manipulate and sort data from a dataset, namely the historical winter olympics medal winners}

\author{Yen}
\keyword{package}
\examples{
			problem1('yen', dnorm, c(-5,5), 100, T)
        	outputlength
          }

