\name{my1rpack}
\alias{my1rpack-package}
\alias{my1rpack}

\docType{package}

\title{
  Homework 2
}

\description{
  Contains a function for generating tikz plots and a dataset containing
  winter Olympics medal statistics by country.
}

\details{
\tabular{ll}{
Package: \tab my1rpack\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-11-01\cr
License: \tab GPL-3\cr
}
}
\author{
  Zhenhan Zhao
  
  Maintainer: Zhenhan Zhao <george.zhao888@gmail.com>

}
\references{
  2012 Fall semester 550400 Homework 2 Description
}

\keyword{ package }

\examples{
library(my1rpack)


}