\name{problem1}
\alias{problem1}
\title{Problem 1 for Homework 2}
\description{Problem 1 for Homework 2. Generates a .tex file and compiles a PDF file with a line plot on it if the user desires. The plot is generated by randomly sampling the specified number of x values over the specified range of numbers and then evaluating the specified mathematical function at each of the x values.}
\usage{problem1(filename, fcnname, range, nsample, compile)}
\arguments{
    \item{filename}{A character vector which contains the user specified name of the output .tex and PDF files.}
    \item{fcnname}{User specified mathematical function which will be   plotted.}
    \item{range}{A numeric vector with two real numbers which specifies the range of numbers which should be randomly sampled from. The order of the two numbers does not matter.}
    \item{nsample}{A numeric vector with a single number which is equal to the size of the sample.}
    \item{compile}{Boolean expression which determines if a PDF file is compiled from the .tex file. If set to TRUE, a PDF file will be compiled. If set to FALSE, no PDF file will be compiled.}
}
\value{
No value is returned.
}
\author{Steven Su <ssu7@jhu.edu>}
\note{Problem 1 for Homework 2.}
\examples{
problem1('bmw1series', dnorm, c(-5,5), 10, TRUE)
problem1('bmw2series', dnorm, c(-3,8), 1000, TRUE)
problem1('bmw3series', dnorm, c(-10,10), 10000, FALSE)
}
