\name{problem1}
\alias{problem1}
\title{Problem 1 for Homework 2}
\description{Problem 1 for Homework 2. Generates LaTeX file and compiles PDF file with a line plot on it. The plot is generated by randomly sampling the specified number of x values over the specicied range of numbers and then evaluating each of the x values using the specified mathematical function.}
\usage{problem1(filename, fcnname, range, nsample, compile)}
\arguments{
    \item{\code{filename}}{A character vector which contains the user specified name of the output tex and PDF files.}
    \item{\code{fcnname}}{User specified mathematical function to which will be   plotted.}
    \item{\code{range}}{A numeric vector with two real numbers which specifies the range of numbers the which should be randomly sampled from.}
    \item{\code{nsample}}{A numeric vector with a single number which is equal to the size of the sample.}
    \item{\code{compile}}{Boolean expression which determines if a PDF file is compiled from the tex file.}
}
\value{
\describe{
    \item{\code{filename.tex}}{A tex file named with the specified filename.}
    \item{\code{filename.pdf}}{A pdf file named with the specified filename.}
    }
}
\author{Steven Su}
\note{None}
\examples{
problem1('bmw1series', dnorm, c(-5,5), 10, TRUE)
problem1('bmw2series', dnorm, c(-3,8), 1000, TRUE)
problem1('bmw3series', dnorm, c(-10,10), 10000, FALSE)
}
