\name{problem2}
\alias{problem2}
\docType{data}
\title{
Problem 2 of Homework 2
}
\description{
Takes data from outside source and makes a table.
}
\usage{problem2}

\format{Data presented in table}

\details{

}
\value{
No value is returned.
} 
\source{\url{http://www.guardian.co.uk/news/datablog/2010/feb/11/winter-olympics-medals-by-country}}




\examples{}

\keyword{ datasets}
