\name{problem2}
\alias{problem2}
\docType{data}
\title{Winter Olympics - Which country has the most medals?}
\description{
The IOC has provided a table of previous winners by year, discipline, gender and country, from 1924 t0 2006}
\usage{
data{problem2}
}
\format{
  A list-mode object with 0 slots each of which contains the overall Bronze, Gold, and Silver medal counts:
  \describe{
  \item{\code{Bronze}}{Good!}
  \item{\code{Gold}}{Awesome!}
  \item{\code{Silver}}{Great!}
}
}
\details{}
\source{\url{http://www.guardian.co.uk/newd/datablog/2010/feb/11/winter-olympics-medals-by-country}}}
\examples{
data(problem2)
}

\keyword{ datasets}
