\name{problem2}
\alias{problem2}
\docType{data}
\title{problem 2}
\description{The function generates statistics of number of medals for each country}
\usage{problem2}
\format{
  A list-mode object ith 45 slots, each of which contains the overall
  Gold, Silver, and Bronze medal counts:
  \describe {
    \item{\code{Gold}}{Awesome!}
    \item{\code{Silver}}{Great!}
    \item{\code{Bronze}}{Good!}
  }
}

\details{}

\source{\url{http://www.guardian.co.uk/news/datablog/2010/feb/11/winter-olympics-medals-by-country}}

\examples{problem2}

\keyword{datasets}