\name{problem2}
\alias{problem2}
\docType{data}
\title{Problem 2 for Homework 2}
\description{Problem 2 for Homework 2. A list-mode object with the number of Bronze, Gold, and Silver medals won by each country over all Winter Olympics from 1924 through 2006.}
\usage{data(problem2)}
\format{
A list-mode object with 45 slots, each of which contains the overall Bronze, Gold, and Silver medal counts for each country:
\describe{
  \item{\code{Bronze}}{Good!}
  \item{\code{Gold}}{Awesome!}
  \item{\code{Silver}}{Great!}
}
}
\details{The included data is for the Winter Olympics of 1924 through 2006.}
\source{\url{http://www.guardian.co.uk/news/datablog/2010/feb/11/winter-olympics-medals-by-country}}
\examples{
data(problem2)
}
\keyword{datasets}
