\name{problem2}
\alias{problem2}
\docType{data}
\title{Problem 2 for Homework 2}
\description{Problem 2 for Homework 2. A list with the number of Bronze, Gold, and Silver medals won by each country over all Winter Olympics from 1924 to 2006.}
\usage{data}
\format{
A list-mode object with 45 slots, each of which contains the overall Bronze, Gold, and Silver medal counts:
\describe{
  \item{\code{Bronze}}{Good!}
  \item{\code{Gold}}{Awesome!}
  \item{\code{Silver}}{Great!}
}
}
\details{The data used is from 1924 through the 2006 Winter Olympics}
\source{\url{http://www.guardian.co.uk/news/datablog/2010/feb/11/winter-olympics-medals-by-country}}
\examples{
load("my1rpack/data/problem2.RData")
}
\keyword{datasets}
