\name{problem2}
\alias{problem2}
\docType{data}
\title{Solution to Problem 2}
\description{This is a workspace showing data of winter olympics medal winners from 1924 until 2006. the functions outputlength and outputtable sorts these medal winners by country and by medal respectively}
\usage{
          outputlength
          outputtable
                 }
\format{
	A list-mode object with 0 slots, each of which contains 
	the overall Broze, Silver, and Gold medal counts:
	\describe{
		\item{\code{Bronze}}{Good!}
		\item{\code{Silver}}{Great!}
		\item{\code{Gold}}{Awesome!}
	}
	
}
\details{}
\source{\url{http://www.guardian.co.uk/news/datablog/2010/feb/11/winter-olympics-medals-by-country}}
\keyword{datasets}
