\name{problem2}
\alias{problem2}
\title{Solution to Problem 2}
\description{This is a workspace showing data of winter olympics medal winners from 1924 until 2006. the functions outputlength and outputtable sorts these medal winners by country and by medal respectively}
\usage{
          outputlength
          outputtable
                 }
\arguments{
  \item{outputlength}{function}
  \item{fcnname}{function}
}

\value{
  \itemize{
    \item nothing returned  
          }
}
\author{Yen
}
\examples{
          outputlength
          }

