\name{problem2}
\alias{problem2}
\docType{data}
\title{Winter Olympics: Which country has won the most medals?}
\description{The IOC has collected and published data of winter olympics medal winners from 1924 until 2006. the functions problem2 sorts these medal winners by country and by medal}
\usage{
          problem2
                 }
\format{
	A list-mode object with 0 slots, each of which contains 
	the overall Broze, Silver, and Gold medal counts:
	\describe{
		\item{\code{Bronze}}{Good!}
		\item{\code{Silver}}{Great!}
		\item{\code{Gold}}{Awesome!}
	}
	
}
\details{}
\source{\url{http://www.guardian.co.uk/news/datablog/2010/feb/11/winter-olympics-medals-by-country}}
\examples{data(problem2)}
\keyword{datasets}
