\name{Problem2}

\alias{Problem2}

\docType{data}

\title{
Function for homework 2 problem 2
}

\description{
  The data containing medal statistics by country.
}

\usage{Problem2}

\format{
  A list-mode object ith 45 slots, each of which contains the overall
  Gold, Silver, and Bronze medal counts:
  \describe{
    \item{\code{Gold}}{Awesome!}
    \item{\code{Silver}}{Great!}
    \item{\code{Bronze}}{Good!}
  }
}

\details{}

\source{\url{http://www.guardian.co.uk/news/datablog/2010/feb/11/winter-olympics-medals-by-country}}

\examples{

}

\keyword{datasets}